Use rvslots format for MediaWiki revisions queries
Per https://phabricator.wikimedia.org/T391900 we need to update the MediaWiki API calls to use the rvslot parameter when fetching revisions, because doing so without the rvslot parameter is long-deprecated and is flooding the Wikipedia logs.

The WikiWho logic for analyzing a series of revisions is in a separate library, and is written to handle revision data in the old format, which is not supported when using the rvslots parameter. Compare:
* https://en.wikipedia.org/w/api.php?action=query&prop=revisions&titles=API|Main%20Page&rvprop=timestamp|user|comment|content
* https://en.wikipedia.org/w/api.php?action=query&prop=revisions&titles=API|Main%20Page&rvslots=*&rvprop=timestamp|user|comment|content

Rather than attempt to update the WikiWho dependency to use the rvslots format, I've implemented a workaround to munge the rvslots results into the old format where '*' (the wikitext of the revision) is a direct property of the revision object.

I tested this locally with on-demand pickles, which are generated with calls to api/handler.py#handle.

There are two other places in the wikiwho_api codebase that make deprecated revisions queries: utils.py#get_latest_revision_timestamps, and the mw_persistence.py test file. I don't think modifying the test is necessary. The timestamps function is only called from an api_editor/manager/commands/fill_notindexed_editor_tables.py function, which I don't think we use. However, since the timeslot is the only rvprop requested, the rvslot format likely doesn't make a difference (and no slots are needed), so if we do need to update that as well, it can likely be done just by adding 'rvslots': '' to the params. I have not made that change, as I don't know what the function is for or how to test it.